% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgpins.gen.R
\name{board}
\alias{board}
\title{Get a package's user-cache pins board}
\usage{
board(pkg, pkg_versioned = TRUE)
}
\arguments{
\item{pkg}{Package name. A character scalar.}

\item{pkg_versioned}{Whether or not to make the board name dependent on \code{pkg}'s version number. If \code{FALSE}, the board is agnostic about \code{pkg}'s version
number. Otherwise, a separate board is created for each \code{pkg} version number (e.g. after an upgrade).}
}
\value{
The user-cache pins board belonging to \code{pkg} (and its version number if \code{pkg_versioned = TRUE}). An object of class
\code{\link[pins:board_folder]{pins_board_folder}}.
}
\description{
Returns a \code{pkg}'s user-cache pins board. The board is automatically created if it doesn't already exist.
}
\seealso{
Other package cache management functions:
\code{\link[=clear_cache]{clear_cache()}},
\code{\link[=ls_cache]{ls_cache()}},
\code{\link[=path_cache]{path_cache()}},
\code{\link[=purge_cache]{purge_cache()}},
\code{\link[=purge_caches]{purge_caches()}}
}
\concept{pkg_cache_mgmt}
