% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgpins.gen.R
\name{boardname}
\alias{boardname}
\title{Get a package's user-cache pins board name}
\usage{
boardname(pkg)
}
\arguments{
\item{pkg}{The package name. A character scalar.}
}
\value{
The board name of the \href{https://pins.rstudio.com/articles/boards-understanding.html}{user-cache pins board} belonging to \code{pkg}, which is at the same
time the name of the filesystem directory beneath \code{\link[=path_cache]{path_cache()}}. A character scalar.
}
\description{
Returns the board name of the \code{pkg}'s user-cache pins board. Note that it is not checked whether the board actually exists on the filesystem or not.
}
\details{
It shouldn't be necessary to rely on this function for the tasks this package is intended to perform, though.
}
