% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgpins-GEN.R
\name{call_to_name}
\alias{call_to_name}
\title{Convert a function call to a (pin) name}
\usage{
call_to_name(
  call = match.call(definition = sys.function(sys.parent(1L)), call =
    sys.call(sys.parent(1L)))
)
}
\arguments{
\item{call}{A \link[base:call]{function call}. Defaults to the call of the calling function.}
}
\value{
A character scalar.
}
\description{
This function strives to create a string that uniquely identifies a function call by the function name and the specified arguments. For example, the
function call \code{foo(a = T, b = "bar", c = 1)} will be converted to \code{"foo_aTRUE-bbar-c1"}.
}
\details{
This function does evaluate object names passed as function arguments. I.e. the function call \code{foo(a = my_var)}, where \code{my_var}'s value is \code{100}, will be
converted to \code{"foo_a100"}. Nevertheless, it should be used with caution.
}
