% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgpins.gen.R
\name{clear}
\alias{clear}
\title{Clear a package's user-cache pins board}
\usage{
clear(pkg, max_age = "1 day")
}
\arguments{
\item{pkg}{The package name. A character scalar.}

\item{max_age}{The age above which cached objects will be deleted. A valid \link[lubridate:as.duration]{lubridate duration}. Defaults to 1 day (24 hours).}
}
\value{
The IDs of the deleted objects invisibly.
}
\description{
Deletes all objects from a package's user-cache pins board that exceed a certain \code{max_age} or were cached with a different version of \code{pkg} than the
currently installed one.
}
\details{
This function could be called on package load/unload, for example.
}
\examples{
\dontrun{
# delete all cached results that are older than 7 days at once on package load:
.onLoad <- function(libname, pkgname) {
  pkgpins::clear(pkg = pkgname,
                 max_age = "7 days")
}}
}
