% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgpins-GEN.R
\name{deregister}
\alias{deregister}
\title{Deregister a package's user-cache pins board}
\usage{
deregister(pkg)
}
\arguments{
\item{pkg}{Package name. A character scalar.}
}
\value{
\code{NULL}.
}
\description{
This function deregisters a package's user-cache pins board.
}
\details{
Ideally, you call this function on \link[base:ns-hooks]{package unload}. See the examples for details.
}
\examples{
\dontrun{
# deregister the cache on package unload (recommended)
.onUnload <- function(libpath) {
  pkgpins::deregister()
}}
}
