% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgpins.gen.R
\name{is_cached}
\alias{is_cached}
\title{Test whether an object exists in a package's user-cache pins board}
\usage{
is_cached(board, id, max_age = "1 day")
}
\arguments{
\item{board}{\link[=board]{Package's user-cache pins board}.}

\item{id}{Pin name uniquely identifying the object to be checked in the \code{pkg}'s user-cache pins board. A character scalar.}

\item{max_age}{Maximum age the cached object is allowed to have. A valid \link[lubridate:as.duration]{lubridate duration}. Defaults to 1 day (24 hours).}
}
\value{
A character scalar, or \code{NULL} if no cached object exists that hasn't exceeded \code{max_age}.
}
\description{
Returns \code{TRUE} if the object is present in the \code{pkg}'s user-cache pins board and hasn't exceeded \code{max_age}, otherwise \code{FALSE}.
}
\seealso{
Other object handling functions:
\code{\link[=cache_obj]{cache_obj()}},
\code{\link[=get_obj]{get_obj()}},
\code{\link[=hash_fn_call]{hash_fn_call()}},
\code{\link[=rm_obj]{rm_obj()}}
}
\concept{obj_handling}
