% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgpins.gen.R
\name{ls_cache}
\alias{ls_cache}
\title{List all objects in a package's user-cache pins board}
\usage{
ls_cache(board)
}
\arguments{
\item{board}{\link[=board]{Package's user-cache pins board}.}
}
\value{
A \link[tibble:tbl_df]{tibble} containing the columns \code{id}, \code{date_time_cached}, \code{file_size} and \code{pins_api_version}.
}
\description{
Lists all objects stored in a \code{pkg}'s user-cache pins board.
}
\seealso{
Other package cache management functions:
\code{\link[=board]{board()}},
\code{\link[=clear_cache]{clear_cache()}},
\code{\link[=path_cache]{path_cache()}},
\code{\link[=purge_cache]{purge_cache()}},
\code{\link[=purge_caches]{purge_caches()}}
}
\concept{pkg_cache_mgmt}
