% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mtcars_metric}
\alias{mtcars_metric}
\title{Motor Trend Car Road Tests, but Metric Consumption}
\format{
\subsection{\code{mtcars_metric}}{

A data frame with 32 observations on 12 (numeric) variables.
\describe{
\item{mpg}{Miles/(US) gallon}
\item{cyl}{Number of cylinders}
\item{disp}{Displacement (cu.in.)}
\item{hp}{Gross horsepower}
\item{drat}{Rear axle ratio}
\item{wt}{Weight (1000 lbs)}
\item{qsec}{1/4 mile time}
\item{vs}{Engine (0 = V-shaped, 1 = straight)}
\item{am}{Transmission (0 = automatic, 1 = manual)}
\item{gear}{Number of forward gears}
\item{carb}{Number of carburetors}
\item{lper100km}{Liters/100 km}
}
}
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
\usage{
mtcars_metric
}
\description{
An update to the classic R dataset: the data was extracted from the 1974
Motor Trend US magazine, and comprises fuel consumption and 10 aspects
of automobile design and performance for 32 automobiles (1973–74 models).
}
\keyword{datasets}
