% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgpins-GEN.R
\name{param_label}
\alias{param_label}
\title{Get predefined parameter label}
\usage{
param_label(name = c("cache_lifespan", "use_cache"))
}
\arguments{
\item{name}{The label name.}
}
\value{
A character scalar.
}
\description{
These are pre-defined labels intended to be used for parameters of the same name in \verb{@param}
\href{https://roxygen2.r-lib.org/articles/rd.html#functions}{roxygen2 tags}.
}
\details{
Recommended usage in \href{https://www.tidyverse.org/blog/2020/03/roxygen2-7-1-0/}{roxygen2 7.1.0} and above:

\verb{@param cache_lifespan `r pkgpins::param_label("cache_lifespan")` }
\verb{@param use_cache `r pkgpins::param_label("use_cache")` }
}
