% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgpins-package.R
\docType{package}
\name{pkgpins-package}
\alias{pkgpins}
\alias{pkgpins-package}
\title{pkgpins: Facilitates Results Caching in R Packages Using the Pins Package}
\description{
pkgpins builds upon the pins package and strives to provide R package authors a simple and standardized way to cache results on the filesystem in a suitable user directory.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pkgpins.rpkg.dev}
  \item \url{https://gitlab.com/rpkg.dev/pkgpins}
  \item Report bugs at \url{https://gitlab.com/rpkg.dev/pkgpins/-/issues}
}

}
\author{
\strong{Maintainer}: Salim Brüggemann \email{salim@rpkg.dev} (\href{https://orcid.org/0000-0002-5329-5987}{ORCID})

}
\keyword{internal}
