% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgpins-GEN.R
\name{register}
\alias{register}
\title{Register a package's user-cache pins board}
\usage{
register(pkg)
}
\arguments{
\item{pkg}{Package name. A character scalar.}
}
\value{
\code{boardname} invisibly if it wasn't registered before, otherwise \code{NULL}.
}
\description{
This function registers a package's user-cache pins board. It will be automatically called when needed.
}
\examples{
\dontrun{
# you could register the cache on package load, although this is not necessary in most cases
.onLoad <- function(libname, pkgname) {
  pkgpins::register(pkgname)
}}
}
