% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgpins.gen.R
\name{rm_obj}
\alias{rm_obj}
\title{Delete a cached object from a package's user-cache pins board}
\usage{
rm_obj(board, id)
}
\arguments{
\item{board}{\link[=board]{Package's user-cache pins board}.}

\item{id}{Pin name uniquely identifying the object to be deleted from the \code{pkg}'s user-cache pins board. A character scalar.}
}
\value{
\code{id}, invisibly.
}
\description{
Removes a cached object from a package's user-cache pins board.
}
\details{
Note that it normally won't be necessary to delete a cached object using this function because \code{\link[=cache_obj]{cache_obj()}} will always overwrite a possibly existing object
in the cache with the same \code{id}. If you want to ensure that no "data corpses" are left behind from using \code{\link[=cache_obj]{cache_obj()}}, consider \link[=clear_cache]{clearing} or
even \link[=purge_cache]{purging} the whole package user-cache pins board at once.
}
\seealso{
Other object handling functions:
\code{\link[=cache_obj]{cache_obj()}},
\code{\link[=get_obj]{get_obj()}},
\code{\link[=hash_fn_call]{hash_fn_call()}},
\code{\link[=is_cached]{is_cached()}}
}
\concept{obj_handling}
