% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgpins-GEN.R
\name{unregister}
\alias{unregister}
\title{Unregister a package's user-cache pins board}
\usage{
unregister(pkg)
}
\arguments{
\item{pkg}{Package name. A character scalar.}
}
\description{
This function unregisters a package's user-cache pins board.
}
\details{
Ideally, you call this function on \link[base:.onUnload]{package unload}. See the examples for details.
}
\examples{
\dontrun{
# register the cache on package unload
.onUnload <- function(libpath) {
  pkgpins::unregister()
}
}
}
