% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgpins.gen.R
\name{board}
\alias{board}
\title{Get a package's user-cache pins board}
\usage{
board(pkg, pkg_version = utils::packageVersion(pkg = pkg))
}
\arguments{
\item{pkg}{Package name. A character scalar.}

\item{pkg_version}{Optional version number of \code{pkg} to include in the board name. A \link[utils:packageDescription]{package version number} or \code{NULL}. If \code{NULL}, the
board is agnostic about \code{pkg}'s version number. Otherwise, a separate board is created for each \code{pkg} version number.}
}
\value{
The user-cache pins board belonging to \code{pkg} and \code{pkg_version}. An object ob class \code{\link[pins:board_folder]{pins_board_folder}}.
}
\description{
Returns a \code{pkg}'s user-cache pins board. The board is automatically created if it doesn't already exist.
}
\seealso{
Other package cache management functions:
\code{\link{clear_cache}()},
\code{\link{ls_cache}()},
\code{\link{path_cache}()},
\code{\link{purge_caches}()},
\code{\link{purge_cache}()}
}
\concept{pkg_cache_mgmt}
