% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgpins-GEN.R
\name{boardname}
\alias{boardname}
\title{Get a package's user-cache pins board name}
\usage{
boardname(pkg)
}
\arguments{
\item{pkg}{Package name. A character scalar.}
}
\value{
The board name of the \href{https://pins.rstudio.com/articles/boards-understanding.html}{user-cache pins board} belonging to \code{pkg}, which is at the same
the name of the filesystem directory beneath \code{\link[pkgpins:path_cache]{pkgpins::path_cache()}}. A character scalar.
}
\description{
This function returns the board name of the \code{pks}'s user-cache pins board.
}
\details{
It shouldn't be necessary to rely on this function for the tasks this package is intended to perform, though.
}
\examples{
\donttest{
pkgpins::boardname(pkg = "not.a.real.pkg")}
}
