% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgpins-GEN.R
\name{call_to_name}
\alias{call_to_name}
\title{Convert the calling function call to a name}
\usage{
call_to_name(
  n_generations_back = 1L,
  add_namespace = TRUE,
  rm_blanks = TRUE,
  sanitize = TRUE,
  non_destructive = FALSE,
  exclude_args = c("use_cache", "cache_lifespan"),
  warn_incomplete = TRUE
)
}
\arguments{
\item{n_generations_back}{The number of generations to go back. See \code{\link[=sys.parent]{sys.parent()}} for details. An integerish scalar.}

\item{add_namespace}{Whether to prefix the generated string with the \href{https://cran.r-project.org/doc/manuals/r-release/R-ints.html#Namespaces}{namespace}
(i.e. package name) the called function belongs to (if any).}

\item{rm_blanks}{Remove all whitespaces from \code{call}'s evaluated argument names.}

\item{sanitize}{Remove characters which are problematic or have special meaning on common filesystems from \code{call}'s evaluated argument names. Enabled by
default. See \code{\link[fs:path_sanitize]{fs::path_sanitize()}} for details about the performed sanitation.}

\item{non_destructive}{Disable all name conversion steps which are known to be destructive (loss of information). See details. This setting implies
\code{santize = FALSE} and \code{rm_blanks = FALSE} and has precedence over them.}

\item{exclude_args}{Argument names to be ignored, i.e. excluded from the generated string. A character vector, or \code{NULL} for no excluded arguments.}

\item{warn_incomplete}{Show a warning if the deparser \emph{suspects} it's unable to properly deparse an argument (might involve false positives). See the
\code{"warnIncomplete"} section in \code{\link[base:deparseOpts]{base::..deparseOpts}} for details. A logical scalar.}
}
\value{
A character scalar.
}
\description{
\ifelse{html}{\figure{lifecycle-deprecated.svg}{options: alt='Deprecated lifecycle'}}{\strong{Deprecated}}

\emph{It's recommended to switch to the more reliable/powerful alternative \code{\link[=call_to_hash]{call_to_hash()}} whenever possible.}

Strives to create a string that uniquely identifies the function call of the calling function by the function's namespace, name and specified arguments. For
example, when called \emph{inside} the function \code{foo()}, it will return
"foo-a=FALSE,b='bar',c=1" when \code{foo()} was called with
\code{foo(a = F, b = "bar", c = 1)}.
}
\details{
This function does evaluate object names passed as function arguments. I.e. the function call \code{foo(a = my_var)}, where \code{my_var}'s value is \code{100}, will be
converted to "foo-a=100".

Nevertheless, it should be used with caution since the name conversion can be destructive: For instance, whitespaces are just deleted, thus
\code{foo(a = "ho ho")} and \code{foo(a = paste("ho", "ho", sep = paste0(rep(" ", 5), collapse = "")))} both produce the name
"foo-a='hoho'" by default. Set \code{rm_blanks = FALSE} in order to keep
whitespaces.

Also, characters which are problematic or have special meaning on common filesystems are removed by default. To prevent this, use \code{sanitize = FALSE}
instead.

To turn off all the known-to-be destructive conversion steps at once, set \code{non_destructive = TRUE}.

A more reliable/powerful alternative to this function is \code{\link[=call_to_hash]{call_to_hash()}} which should be preferred whenever possible.
}
\examples{
# By default, the parent function call is returned ...
foo <- function(a, b, c) pkgpins::call_to_name()
foo("ya", "hoo")

# ... but you can go back further and return e.g. the grandparent function call
foo <- function(a) pkgpins::call_to_name(n_generations_back = 2L)
bar <- function(b) foo()
bar()

# Whitespaces are removed by default
# which means the following produces identical names ...
foo <- function(a, b, c) pkgpins::call_to_name()
foo(4 - 2, 'a \" r', list("1, 77"))
foo(1+1, 'ar', list('1,77'))

# ... to avoid this and produce distinct names instead, set `rm_blanks = FALSE` ...
foo <- function(a, b, c) pkgpins::call_to_name(rm_blanks = FALSE)
foo(4 - 2, 'a \" r', list("1, 77"))
foo(1+1, 'ar', list('1,77'))

# ... or just disable the known to be destructive conversion steps altogether
foo <- function(a, b, c) pkgpins::call_to_name(non_destructive = TRUE)
foo(4 - 2, 'a \" r', list("1, 77"))
foo(1+1, 'ar', list('1,77'))

# Arguments can be excluded by name
foo <- function(a, b, c) pkgpins::call_to_name(exclude_args = c("a", "c"))
foo(1+1, 'ar', list('1,77'))

\dontrun{
# R's deparsing warns about possible incompletion when arguments are
# formulas or curly-braced function definitions ...
foo <- function(a, b, c) pkgpins::call_to_name()
foo(~ 1)
foo(function(x){x})

# ... to mute those warnings, set `warn_incomplete = FALSE`
foo <- function(a, b, c) pkgpins::call_to_name(warn_incomplete = FALSE)
foo(~ 1)
foo(function(x){x})}
}
