% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgpins-GEN.R
\name{clear}
\alias{clear}
\title{Delete all objects from a package's user-cache pins board exceeding a certain age}
\usage{
clear(pkg, max_age = "1 day")
}
\arguments{
\item{pkg}{Package name. A character scalar.}

\item{max_age}{The age above which cached objects will be deleted. A valid \link[lubridate:as.duration]{lubridate duration}. Defaults to 1 day (24 hours).}
}
\value{
The IDs of the deleted objects invisibly.
}
\description{
This function allows to clean up old caching left-overs. It could be called on package load/unload, for example.
}
\examples{
\dontrun{
# delete all cached results that are older than 7 days at once on package load:
.onLoad <- function(libname, pkgname) {
  pkgpins::clear(max_age = "7 days")
}}
}
