% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgpins.gen.R
\name{get_obj}
\alias{get_obj}
\title{Get a cached object from a package's user-cache pins board}
\usage{
get_obj(board, id, max_age = "1 day")
}
\arguments{
\item{board}{\link[=board]{Package's user-cache pins board}.}

\item{id}{Pin name uniquely identifying the object to be retrieved from the \code{pkg}'s user-cache pins board. A character scalar.}

\item{max_age}{Maximum age the cached object is allowed to have. A valid \link[lubridate:as.duration]{lubridate duration}. Defaults to 1 day (24 hours).}
}
\value{
The cached object if it is not older than \code{max_age}, otherwise \code{NULL}.
}
\description{
Retrieves a cached object from a package's user-cache pins board \emph{if} it is not older than \code{max_age}.
}
\seealso{
Other object handling functions:
\code{\link{cache_obj}()},
\code{\link{hash_fn_call}()},
\code{\link{is_cached}()},
\code{\link{rm_obj}()}
}
\concept{obj_handling}
