% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgpins-GEN.R
\name{hash_fn_call}
\alias{hash_fn_call}
\title{Hash a function call}
\usage{
hash_fn_call(pkg, from_fn, ...)
}
\arguments{
\item{pkg}{The namespace \code{from_fn} belongs to, typically a package name. A character scalar.}

\item{from_fn}{The name of the function from which \code{hash_fn_call()} is called. A character scalar.}

\item{...}{The arguments \code{from_fn} was called with. Any arguments omitted here won't be taken into account when generating the hash.}
}
\value{
A character scalar.
}
\description{
Creates a string that uniquely identifies a function call by the function's namespace, name and a hash of the specified arguments.
}
\details{
This function does evaluate object names passed as function arguments. I.e. the function call \code{mypkg::foo(a = my_var)} will be cached as
"mypkg-foo-ff6a1ba723cae45aeee265a71493cecf" if \code{my_var}'s value is \code{100}, whereas it will become "mypkg-foo-3f6860362d51e26829cc6f2f3b2968e9" if \code{my_var}'s value is \code{101}. See examples
below.
}
\examples{
foo <- function(a, b = "default") pkgpins::hash_fn_call("mypkg", "foo", a, b)

foo("bar")
a <- "bar"
foo(a)
a <- "barbie"
foo(a)
a |> foo()
a |> foo(b = "default")

# you can also create a hash that is based on a subset of all function arguments
foo <- function(a, b = "default") pkgpins::hash_fn_call("mypkg", "foo", b)

foo("bar")
foo("barrr")
foo(b = "bar")
}
