% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgpins-GEN.R
\name{is_cached}
\alias{is_cached}
\title{Test whether an object exists in a package's user-cache pins board}
\usage{
is_cached(id, pkg, max_age = "1 day")
}
\arguments{
\item{id}{The pin name uniquely identifying the object to be checked in the \code{pkg}'s user-cache pins board. A character scalar.}

\item{pkg}{The package name. A character scalar.}

\item{max_age}{The maximum age the cached object is allowed to have. A valid \link[lubridate:as.duration]{lubridate duration}. Defaults to 1 day (24 hours).}
}
\value{
A character scalar, or \code{NULL} if no cached object exists that hasn't exceeded \code{max_age}.
}
\description{
Returns \code{TRUE} if the object is present in the \code{pkg}'s user-cache pins board and hasn't exceeded \code{max_age}, otherwise \code{FALSE}.
}
