% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgpins-GEN.R
\name{ls_cache}
\alias{ls_cache}
\title{List all objects in a package's user-cache pins board}
\usage{
ls_cache(pkg)
}
\arguments{
\item{pkg}{Package name. A character scalar.}
}
\value{
A \link[tibble:tbl_df]{tibble}.
}
\description{
This function lists all object \code{id}s and the date and time they were \code{cached}.
}
