% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgpins-GEN.R
\name{ls_cache}
\alias{ls_cache}
\title{List all objects in a package's user-cache pins board}
\usage{
ls_cache(pkg)
}
\arguments{
\item{pkg}{The package name. A character scalar.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Lists all object \code{id}s belonging to a \code{pkg}'s user-cache pins board, together with some metadata (see \code{\link[=cache_obj]{cache_obj()}} for details).
}
