% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgpins.gen.R
\name{path_cache}
\alias{path_cache}
\title{Get a package's user-cache pins board path}
\usage{
path_cache(board)
}
\arguments{
\item{board}{\link[=board]{Package's user-cache pins board}.}
}
\value{
A \link[fs:fs_path]{path}.
}
\description{
Returns the filesystem path to the \code{pkg}'s user-cache pins board. Note that this path depends on the operating system this function is called from.
}
\details{
It shouldn't be necessary to rely on this function for the tasks this package is intended to perform, though.
}
\seealso{
Other package cache management functions:
\code{\link{board}()},
\code{\link{clear_cache}()},
\code{\link{ls_cache}()},
\code{\link{purge_cache}()},
\code{\link{purge_caches}()}
}
\concept{pkg_cache_mgmt}
