% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgpins.gen.R
\name{purge_cache}
\alias{purge_cache}
\title{Purge a package's user-cache pins board}
\usage{
purge_cache(board)
}
\arguments{
\item{board}{\link[=board]{Package's user-cache pins board}.}
}
\description{
Deletes \emph{all} objects from a package's user-cache pins board. All boards belonging to the same package as \code{board} are deleted, too.
}
