% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgpins.gen.R
\name{purge_cache}
\alias{purge_cache}
\title{Purge a package's user-cache pins board}
\usage{
purge_cache(pkg)
}
\arguments{
\item{pkg}{The package name. A character scalar.}
}
\value{
The deleted filesystem \link[fs:fs_path]{path}, invisibly. Of length 0 if the package's user-cache pins board didn't exist.
}
\description{
Deletes \emph{all} objects from a package's user-cache pins board.
}
