% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgpins.gen.R
\name{purge_caches}
\alias{purge_caches}
\title{Purge all package user-cache pins boards}
\usage{
purge_caches()
}
\value{
The deleted filesystem \link[fs:fs_path]{paths}, invisibly. Of length 0 if no package user-cache pins boards did exist.
}
\description{
Deletes the user-cache pins board of \emph{all} packages.
}
