% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgpins-GEN.R
\name{rm_obj}
\alias{rm_obj}
\title{Delete a cached object from a package's user-cache pins board}
\usage{
rm_obj(id, pkg)
}
\arguments{
\item{id}{The pin name uniquely identifying the object to be deleted from the \code{pkg}'s user-cache pins board. A character scalar.}

\item{pkg}{The package name. A character scalar.}
}
\value{
\code{id}, invisibly.
}
\description{
Removes a cached object from a package's user-cache pins board.
}
\details{
Note that it normally won't be necessary to delete a cached object using this function because \code{\link[=cache_obj]{cache_obj()}} will always overwrite a possibly existing object
in the cache with the same \code{id}. If you want to ensure that no "data corpses" are left behind from using \code{\link[=cache_obj]{cache_obj()}}, consider clearing the whole package
user-cache pins board at once using \code{\link[=clear]{clear()}}.
}
